+++ XVim/XVimTextObjectEvaluator.m
@@ -52,7 +52,7 @@
     [super didEndHandler];
 }
 
-- (float)insertionPointHeightRatio{
+- (CGFloat)insertionPointHeightRatio{
     return 0.5;
 }
